+++ src/backend/index.js
@@ -6,7 +6,8 @@ const fs = require('fs')
 const jwt = require('express-jwt')
 const jwksRsa = require('jwks-rsa')
 const morgan = require('morgan')
-const routes = require('./routes')
+const api = require('./api')
+const pql = require('./pql')
 
 const app = express()
 const port = 1337
@@ -66,7 +67,20 @@ app.use(
       stream: accessLogStream,
     }
   ), 
-  routes
+  api
+)
+
+app.use(
+  '/pql',
+  checkJwt,
+  (err, req, res, next) => {
+    if (err.name === 'UnauthorizedError') {
+      res.status(401).json(err)
+    } else {
+      next()
+    }
+  },
+  pql
 )
 
 app.listen(port, () => console.log(`PHOENIX ONLINE. PORT ${port}!`))