+++ src/backend/pql/getAggPipelines/parser/parser.js
@@ -1,0 +1,3 @@
+const regex = /(\w*) = \(?\w*[,-\w]*\)?/gi
+
+module.exports = pql => pql.match(regex)